Christian

+++ TimePunch/Assets/Scripts'/Rotation.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Rotation : MonoBehaviour {
-    private float rpm = 10.0;
+    private float rpm = 10.0f;
 	// Use this for initialization
 	void Start () {
 		
@@ -11,6 +11,6 @@ public class Rotation : MonoBehaviour {
 
 	// Update is called once per frame
 	void Update () {
-        transform.Rotate(0, 6.0 * rpm * Time.deltaTime, 0);
+        transform.Rotate(0, (float)(6.0 * rpm * Time.deltaTime), 0);
 	}
 }
